refactor: move the Session title helpers to their only owner
With naming rehomed to the Host, `@maka/runtime/session-title` had no consumer
left inside the Runtime: the prompt, the cleaner, the timeout, the source
extractor and the fallback are read only by the Host's Session-effect
coordinator and its model authority. The module was the last residue of the
Runtime's old half of the effect.

It now sits beside them in `@maka/runtime-host`, and `@maka/runtime` drops the
`./session-title` export subpath. Nothing publishes these packages, so a
subpath with no importer in the repository has no other demand to serve.

Generated-by: Claude Code
